+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Unreleased
 
+- Fixed bugs in `Transaction.getSignatures()`
+- Added `ChunkedTransaction.getAllSignatures()`
 - Fixed bug in `TopicMessageQuery`
 - Fixed bugs in `ChunkedTransaction.getAllTransactionHashesPerNode()`
 - Setters and getters of `List` fields (such as `TokenBurnTransaction.setSerials()`) now ensure that the user can't gain mutable access to the List field in the Transaction or Query.